[3DS] proper screen output
I'm putting this on-hold for performance reasons and moving onto PicoDrive.

What needs to be done:
     * Framebuffer should be scaled by GPU DMA engine using a GX_* command.
     * Framebuffer format should be set to RGB565, to avoid stupid conversion.
     * Menu, listing files on sdcard.
     * Input handling, currently input_update_() is pretty much empty.

This commit will run Sonic 1 a little bit.

This will not run on ninjhax right now due to a limitation, which will be
fixed in the next version.

It will run (horribly slow) in the 3dmoo emulator.